==> paper_I/1_code/0_Kinematics_Table.py <==
import csvfrom astropy.io import fitsimport numpy as npfrom pathlib import Pathfrom astropy.io import ascii#%precision 2	#print(GalaxyList[0].keys()) #the columns' namesdef import_sigma(index):		dictS = {}	dictE = {}	fileS = "../0_data/SAMI/fcc" + str(int(index)) + "_blue_results_simple_onebin.fits"	fileE = "../0_data/SAMI/fcc" + str(int(index)) + "_blue_results_mc_onebin.fits"	my_file = Path(fileS)	if my_file.exists():		hdulist = fits.open(fileS)		table = hdulist[1].data		sigma = np.mean(table["SPXF"])		dictS.update({"Sigma": "%.4f" % sigma})		velocity = np.mean(table["VPXF"])		dictS.update({"Velocity": "%.4f" % velocity})		h3 = np.mean(table["H3PXF"])		dictS.update({"H3": "%.4f" % h3})		h4 = np.mean(table["H4PXF"])		dictS.update({"H4": "%.4f" % h4})		else:		dictS.update({dictkey: "-"})			my_file = Path(fileE)	if my_file.exists():		hdulist = fits.open(fileE)		table = hdulist[1].data		Serror= np.mean(table['DSp'])		dictE.update({"Error.Sigma": "%.4f" % Serror})		Verror= np.mean(table['DVp'])		dictE.update({"Error.Velocity": "%.4f" % Verror})		h3error = np.mean(table["Dh3p"])		dictE.update({"Error.H3": "%.4f" % h3error})		h4error = np.mean(table["Dh4p"])		dictE.update({"Error.H4": "%.4f" % h4error})		else:		dictE.update({dictkey: "-"})				return (dictS, dictE)with open('../2_pipeline/0_Galaxies_Table/Galaxies_Table.csv') as csvfile:	listdict_info = list(csv.DictReader(csvfile))ListDictGal = []for d in listdict_info:	#dictgal = {'FCC':d['FCC...'], 'R_e(arcsec)':d['R_e']}	dictgal = {'FCC':d['FCC']}	dictgal.update(import_sigma(d['FCC'])[0])	dictgal.update(import_sigma(d['FCC'])[1])	ListDictGal.append(dictgal)keys = ListDictGal[0].keys()with open('../2_pipeline/0_Kinematics_Table/Kinematics_Table.csv', 'w') as output_file:    dict_writer = csv.DictWriter(output_file, keys)    dict_writer.writeheader()    dict_writer.writerows(ListDictGal)
